refactoring create event page for readability
